리팩토링 완료, 테스트 수정
