+++ src/main/java/com/skka/adapter/common/exception/BadRequestException.java
@@ -1,0 +1,14 @@
+package com.skka.adapter.common.exception;
+
+import lombok.Getter;
+
+@Getter
+public class BadRequestException extends RuntimeException {
+
+    private final int code;
+
+    public BadRequestException(final ErrorType errorType) {
+        super(errorType.getMessage());
+        this.code = errorType.getCode();
+    }
+}
