+++ src/main/java/com/skka/adapter/inbound/api/studyseat/webrequest/CommandCheckoutScheduleWebRequestV1.java
@@ -1,0 +1,13 @@
+package com.skka.adapter.inbound.api.studyseat.webrequest;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class CommandCheckoutScheduleWebRequestV1 {
+
+    private String scheduleState;
+}
