+++ src/main/java/com/skka/adapter/outbound/jpa/customer/CustomerJpaRepository.java
@@ -1,0 +1,7 @@
+package com.skka.adapter.outbound.jpa.customer;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CustomerJpaRepository extends JpaRepository<CustomerEntity, Long> {
+
+}
