+++ src/main/java/com/skka/adapter/outbound/jpa/studyseat/StudySeatJpaRepository.java
@@ -1,0 +1,7 @@
+package com.skka.adapter.outbound.jpa.studyseat;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface StudySeatJpaRepository extends JpaRepository<StudySeatEntity, Long> {
+
+}
