--- src/main/java/com/skka/adaptor/inbound/api/studyseat/webrequest/CommandChangeStudyTimeWebRequestV1.java
@@ -1,20 +1,0 @@
-package com.skka.adaptor.inbound.api.studyseat.webrequest;
-
-import java.time.LocalDateTime;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.ToString;
-import org.springframework.format.annotation.DateTimeFormat;
-
-@Getter
-@RequiredArgsConstructor
-@ToString
-public class CommandChangeStudyTimeWebRequestV1 {
-    private final long customerId;
-
-    @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-    private final LocalDateTime changingStartedTime;
-
-    @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-    private final LocalDateTime changingEndTime;
-}
