--- src/main/java/com/skka/adaptor/inbound/api/studyseat/webrequest/CommandCheckoutScheduleWebRequestV1.java
@@ -1,11 +1,0 @@
-package com.skka.adaptor.inbound.api.studyseat.webrequest;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class CommandCheckoutScheduleWebRequestV1 {
-
-    private final String scheduleState;
-}
