--- src/main/java/com/skka/adaptor/inbound/api/studyseat/webrequest/CommandReserveSeatWebRequestV1.java
@@ -1,20 +1,0 @@
-package com.skka.adaptor.inbound.api.studyseat.webrequest;
-
-import java.time.LocalDateTime;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.ToString;
-import org.springframework.format.annotation.DateTimeFormat;
-
-@Getter
-@RequiredArgsConstructor
-@ToString
-public class CommandReserveSeatWebRequestV1 {
-    private final long customerId;
-
-    @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm")
-    private final LocalDateTime startedTime;
-
-    @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm")
-    private final LocalDateTime endTime;
-}
