--- src/main/java/com/skka/adaptor/outbound/jpa/BaseEntity.java
@@ -1,32 +1,0 @@
-package com.skka.adaptor.outbound.jpa;
-
-import java.time.LocalDateTime;
-import javax.persistence.Column;
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@EntityListeners(AuditingEntityListener.class)
-@MappedSuperclass
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-public abstract class BaseEntity {
-
-    @Column(nullable = false, updatable = false, name = "created_at")
-    @CreatedDate
-    private LocalDateTime createdAt;
-
-    @Column(nullable = false, name = "last_modified")
-    @LastModifiedDate
-    private LocalDateTime lastModified;
-
-    protected BaseEntity(LocalDateTime createdAt, LocalDateTime lastModified) {
-        this.createdAt = createdAt;
-        this.lastModified = lastModified;
-    }
-}
