--- src/main/java/com/skka/adaptor/outbound/jpa/studyseat/repository/StudySeatJpaRepository.java
@@ -1,8 +1,0 @@
-package com.skka.adaptor.outbound.jpa.studyseat.repository;
-
-import com.skka.adaptor.outbound.jpa.studyseat.StudySeatEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StudySeatJpaRepository extends JpaRepository<StudySeatEntity, Long> {
-
-}
