+++ src/main/java/com/skka/domain/customer/Customer.java
@@ -1,12 +1,12 @@
 package com.skka.domain.customer;
 
-import static com.skka.adaptor.common.exception.ErrorType.INVALID_CUSTOMER_EMAIL;
-import static com.skka.adaptor.common.exception.ErrorType.INVALID_CUSTOMER_NAME;
-import static com.skka.adaptor.common.exception.ErrorType.INVALID_CUSTOMER_TEL;
-import static com.skka.adaptor.util.Util.require;
+import static com.skka.adapter.common.exception.ErrorType.INVALID_CUSTOMER_EMAIL;
+import static com.skka.adapter.common.exception.ErrorType.INVALID_CUSTOMER_NAME;
+import static com.skka.adapter.common.exception.ErrorType.INVALID_CUSTOMER_TEL;
+import static com.skka.adapter.util.Util.require;
 
-import com.skka.adaptor.outbound.jpa.customer.CustomerEntity;
-import com.skka.adaptor.outbound.jpa.studyseat.schedule.ScheduleEntity;
+import com.skka.adapter.outbound.jpa.customer.CustomerEntity;
+import com.skka.adapter.outbound.jpa.studyseat.schedule.ScheduleEntity;
 import com.skka.domain.studyseat.schedule.Schedule;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -26,37 +26,25 @@ public class Customer {
     private String password;
     private String tel;
     private List<Schedule> schedules = new ArrayList<>();
-//
-//    private LocalDateTime lastModified;
-//    private LocalDateTime createdAt;
+    private LocalDateTime lastModified;
+    private LocalDateTime createdAt;
 
     private Customer(
         final long id,
         final String name,
         final String email,
         final String password,
-        final String tel
+        final String tel,
+        final LocalDateTime lastModified,
+        final LocalDateTime createdAt
     ) {
         this.id = id;
         this.name = name;
         this.email = email;
         this.password = password;
         this.tel = tel;
-    }
-
-    public static Customer of(
-        final long id,
-        final String name,
-        final String email,
-        final String password,
-        final String tel
-    ) {
-
-        require(o -> name == null, name, INVALID_CUSTOMER_NAME);
-        require(o -> email == null, email, INVALID_CUSTOMER_EMAIL);
-        require(o -> tel == null, tel, INVALID_CUSTOMER_TEL);
-
-        return new Customer(id, name, email, password, tel);
+        this.lastModified = lastModified;
+        this.createdAt = createdAt;
     }
 
     public static Customer of(
@@ -65,23 +53,30 @@ public class Customer {
         final String email,
         final String password,
         final String tel,
-        final List<ScheduleEntity> scheduleEntityList
+        final LocalDateTime lastModified,
+        final LocalDateTime createdAt
     ) {
 
         require(o -> name == null, name, INVALID_CUSTOMER_NAME);
         require(o -> email == null, email, INVALID_CUSTOMER_EMAIL);
         require(o -> tel == null, tel, INVALID_CUSTOMER_TEL);
 
-        return new Customer(id, name, email, password, tel);
+        return new Customer(id, name, email, password, tel, lastModified, createdAt);
     }
 
     public CustomerEntity toCustomerEntity() {
-        CustomerEntity a = CustomerEntity.of(id, name, email, password, tel);
-        System.out.println("222 = " + a);
-        return a;
-    }
+        CustomerEntity customerEntity = CustomerEntity.of(id, name, email, password, tel);
+        schedules.forEach(scheduleDomain -> customerEntity.getSchedules().add(
+            ScheduleEntity.of(
+                scheduleDomain.getId(),
+                scheduleDomain.getCustomer().toCustomerEntity(),
+                scheduleDomain.getStudySeat().toStudySeatEntity(),
+                scheduleDomain.getStartedTime(),
+                scheduleDomain.getEndTime(),
+                scheduleDomain.getState()
+            )
+        ));
 
-    public void setSchedules(List<ScheduleEntity> schedule) {
-//        schedules.add(schedule.toSchedule());
+        return customerEntity;
     }
 }
