+++ src/test/java/com/skka/application/studyseat/StudySeatServiceTest.java
@@ -64,10 +64,10 @@ class StudySeatServiceTest {
 
         // when
         when(customerRepository.findById(command.getCustomerId()))
-            .thenReturn(Optional.ofNullable(CUSTOMER));
+            .thenReturn(CUSTOMER);
 
         when(studySeatRepository.findById(studySeatId))
-            .thenReturn(Optional.ofNullable(STUDY_SEAT));
+            .thenReturn(STUDY_SEAT);
 
         // then
         CommandReserveSeatResponse actual = studySeatService.reserveSeat(command, studySeatId);
@@ -92,10 +92,10 @@ class StudySeatServiceTest {
             long studySeatId = 1L;
 
             when(customerRepository.findById(command.getCustomerId()))
-                .thenReturn(Optional.ofNullable(CUSTOMER));
+                .thenReturn(CUSTOMER);
 
             when(studySeatRepository.findById(studySeatId))
-                .thenReturn(Optional.ofNullable(STUDY_SEAT));
+                .thenReturn(STUDY_SEAT);
 
             assertThrows(IllegalStateException.class,
                 () -> studySeatService.reserveSeat(command, studySeatId),
@@ -115,10 +115,10 @@ class StudySeatServiceTest {
             long studySeatId = 1L;
 
             when(customerRepository.findById(command.getCustomerId()))
-                .thenReturn(Optional.ofNullable(CUSTOMER));
+                .thenReturn(CUSTOMER);
 
             when(studySeatRepository.findById(studySeatId))
-                .thenReturn(Optional.ofNullable(STUDY_SEAT));
+                .thenReturn(STUDY_SEAT);
 
             assertThrows(IllegalStateException.class,
                 () -> studySeatService.reserveSeat(command, studySeatId),
@@ -138,10 +138,10 @@ class StudySeatServiceTest {
             long studySeatId = 1L;
 
             when(customerRepository.findById(command.getCustomerId()))
-                .thenReturn(Optional.ofNullable(CUSTOMER));
+                .thenReturn(CUSTOMER);
 
             when(studySeatRepository.findById(studySeatId))
-                .thenReturn(Optional.ofNullable(STUDY_SEAT));
+                .thenReturn(STUDY_SEAT);
 
             assertThrows(IllegalStateException.class,
                 () -> studySeatService.reserveSeat(command, studySeatId),
@@ -161,10 +161,10 @@ class StudySeatServiceTest {
             long studySeatId = 1L;
 
             when(customerRepository.findById(command.getCustomerId()))
-                .thenReturn(Optional.ofNullable(CUSTOMER));
+                .thenReturn(CUSTOMER);
 
             when(studySeatRepository.findById(studySeatId))
-                .thenReturn(Optional.ofNullable(STUDY_SEAT));
+                .thenReturn(STUDY_SEAT);
 
             assertThrows(IllegalStateException.class,
                 () -> studySeatService.reserveSeat(command, studySeatId),
@@ -184,10 +184,10 @@ class StudySeatServiceTest {
             long studySeatId = 1L;
 
             when(customerRepository.findById(command.getCustomerId()))
-                .thenReturn(Optional.ofNullable(CUSTOMER));
+                .thenReturn(CUSTOMER);
 
             when(studySeatRepository.findById(studySeatId))
-                .thenReturn(Optional.ofNullable(STUDY_SEAT));
+                .thenReturn(STUDY_SEAT);
 
             assertThrows(IllegalStateException.class,
                 () -> studySeatService.reserveSeat(command, studySeatId),
@@ -207,10 +207,10 @@ class StudySeatServiceTest {
             long studySeatId = 1L;
 
             when(customerRepository.findById(command.getCustomerId()))
-                .thenReturn(Optional.ofNullable(CUSTOMER));
+                .thenReturn(CUSTOMER);
 
             when(studySeatRepository.findById(studySeatId))
-                .thenReturn(Optional.ofNullable(STUDY_SEAT));
+                .thenReturn(STUDY_SEAT);
 
             assertThrows(IllegalStateException.class,
                 () -> studySeatService.reserveSeat(command, studySeatId),
@@ -230,10 +230,10 @@ class StudySeatServiceTest {
             long studySeatId = 1L;
 
             when(customerRepository.findById(command.getCustomerId()))
-                .thenReturn(Optional.ofNullable(CUSTOMER));
+                .thenReturn(CUSTOMER);
 
             when(studySeatRepository.findById(studySeatId))
-                .thenReturn(Optional.ofNullable(STUDY_SEAT));
+                .thenReturn(STUDY_SEAT);
 
             assertThrows(IllegalStateException.class,
                 () -> studySeatService.reserveSeat(command, studySeatId),
@@ -253,10 +253,10 @@ class StudySeatServiceTest {
             long studySeatId = 1L;
 
             when(customerRepository.findById(command.getCustomerId()))
-                .thenReturn(Optional.ofNullable(CUSTOMER));
+                .thenReturn(CUSTOMER);
 
             when(studySeatRepository.findById(studySeatId))
-                .thenReturn(Optional.ofNullable(STUDY_SEAT));
+                .thenReturn(STUDY_SEAT);
 
             assertThrows(IllegalStateException.class,
                 () -> studySeatService.reserveSeat(command, studySeatId),
@@ -276,10 +276,10 @@ class StudySeatServiceTest {
             long studySeatId = 1L;
 
             when(customerRepository.findById(command.getCustomerId()))
-                .thenReturn(Optional.ofNullable(CUSTOMER));
+                .thenReturn(CUSTOMER);
 
             when(studySeatRepository.findById(studySeatId))
-                .thenReturn(Optional.ofNullable(STUDY_SEAT));
+                .thenReturn(STUDY_SEAT);
 
             assertThrows(IllegalStateException.class,
                 () -> studySeatService.reserveSeat(command, studySeatId),
@@ -297,7 +297,7 @@ class StudySeatServiceTest {
 
         // when
         when(studySeatRepository.findById(extractingStudySeatId))
-            .thenReturn(Optional.ofNullable(STUDY_SEAT));
+            .thenReturn(STUDY_SEAT);
 
         // then
         CommandExtractScheduleResponse actual = studySeatService.extractSchedule(extractingStudySeatId, scheduleId);
@@ -318,7 +318,7 @@ class StudySeatServiceTest {
 
         // when
         when(studySeatRepository.findById(studySeatId))
-            .thenReturn(Optional.ofNullable(STUDY_SEAT));
+            .thenReturn(STUDY_SEAT);
 
         assertThrows(IllegalStateException.class,
             () -> studySeatService.extractSchedule(
@@ -343,10 +343,10 @@ class StudySeatServiceTest {
 
         // when
         when(studySeatRepository.findById(studySeatId))
-            .thenReturn(Optional.ofNullable(STUDY_SEAT));
+            .thenReturn(STUDY_SEAT);
 
         when(customerRepository.findById(command.getCustomerId()))
-            .thenReturn(Optional.ofNullable(CUSTOMER));
+            .thenReturn(CUSTOMER);
 
         // then
         CommandChangeStudyTimeResponse actual = studySeatService.changeStudyTime(
@@ -379,10 +379,10 @@ class StudySeatServiceTest {
 
         // when
         when(studySeatRepository.findById(studySeatId))
-            .thenReturn(Optional.ofNullable(STUDY_SEAT));
+            .thenReturn(STUDY_SEAT);
 
         when(customerRepository.findById(command.getCustomerId()))
-            .thenReturn(Optional.ofNullable(CUSTOMER));
+            .thenReturn(CUSTOMER);
 
         // then
         assertThrows(IllegalArgumentException.class,
@@ -410,10 +410,10 @@ class StudySeatServiceTest {
 
         // when
         when(studySeatRepository.findById(studySeatId))
-            .thenReturn(Optional.ofNullable(STUDY_SEAT));
+            .thenReturn(STUDY_SEAT);
 
         when(customerRepository.findById(command.getCustomerId()))
-            .thenReturn(Optional.ofNullable(CUSTOMER));
+            .thenReturn(CUSTOMER);
 
         assertThrows(IllegalStateException.class,
             () -> studySeatService.changeStudyTime(
@@ -438,7 +438,7 @@ class StudySeatServiceTest {
 
         // when
         when(studySeatRepository.findById(studySeatId))
-            .thenReturn(Optional.ofNullable(STUDY_SEAT));
+            .thenReturn(STUDY_SEAT);
 
         // then
         CommandCheckOutScheduleResponse actual = studySeatService.checkoutSchedule(
@@ -467,7 +467,7 @@ class StudySeatServiceTest {
 
         // when
         when(studySeatRepository.findById(studySeatId))
-            .thenReturn(Optional.ofNullable(STUDY_SEAT));
+            .thenReturn(STUDY_SEAT);
 
         // then
         assertThrows(IllegalStateException.class,
@@ -494,7 +494,7 @@ class StudySeatServiceTest {
 
         // when
         when(studySeatRepository.findById(studySeatId))
-            .thenReturn(Optional.ofNullable(STUDY_SEAT));
+            .thenReturn(STUDY_SEAT);
 
         // then
         assertThrows(IllegalStateException.class,
