+++ src/testFixtures/java/com/skka/customer/CustomerFixture.java
@@ -1,6 +1,7 @@
 package com.skka.customer;
 
 import com.skka.domain.customer.Customer;
+import java.time.LocalDateTime;
 
 public class CustomerFixture {
 
@@ -9,6 +10,8 @@ public class CustomerFixture {
         "용용이",
         "yongyong@naver.com",
         "aaaaaaa",
-        "010-1231-1727"
+        "010-1231-1727",
+        LocalDateTime.of(2023, 1, 10, 12, 10),
+        LocalDateTime.of(2023, 1, 10, 15, 10)
     );
 }
